Fix browser navigation for nav items
This was partly broken before introducing bookmarkable urls for events,
and with that got more visible. E.g. this was now unexpected behavior:
* Open https://e4f-hamburg.de/#event-06-12-23-e4f-vernetzungstreffen
* Close the event / modal dialog
* Click on "Kontakt"
  * scrolls as expected, but doesn't change the url hash (bug)
* Click browser back
  * this now opens the event / modal dialog again (unexpected - caused
    by the previous issue, that clicking on "Kontakt" didn't change the
    url

This fix is now based on
https://github.com/kswedberg/jquery-smooth-scroll?tab=readme-ov-file#fnsmoothscroll
and the mentioned back button support as shown in https://github.com/kswedberg/jquery-smooth-scroll/blob/master/demo/hashchange.html

The same issue is described in
https://github.com/mmistakes/minimal-mistakes/issues/1767, which is
related because our agency plugin also uses smooth scrolling.